更新表结构
