+++ model/like.go
@@ -1,0 +1,10 @@
+package model
+
+type Like struct {
+	UserID  uint
+	VideoID uint
+}
+
+func (Like) TableName() string {
+	return "like"
+}
